requirements.txt
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
confusable-homoglyphs==3.2.0
defusedxml==0.6.0
Django==2.2.3
django-allauth==0.39.1
django-crispy-forms==1.7.2
django-registration==3.0.1
django-rest==0.0.1
django-rest-auth==0.9.5
django-rest-framework==0.1.0
django-webpack-loader==0.6.0
djangorestframework==3.10.2
idna==2.8
misaka==2.1.1
mysqlclient==1.4.2
oauthlib==3.0.2
Pillow==6.1.0
pycparser==2.19
python3-openid==3.1.0
pytz==2019.1
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.3
